requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
beautifulsoup4==4.9.3
brotlipy==0.7.0
chardet==4.0.0
click==8.1.3
cymruwhois==1.6
dnspython==2.2.1
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
geoip2==4.6.0
googlesearch-python==1.1.0
idna==2.10
itsdangerous==2.1.2
joblib==1.2.0
maxminddb==2.2.0
multidict==6.0.2
numpy==1.23.4
pandas==1.5.1
python-whois==0.8.0
pywin32==302
requests==2.25.1
scikit-learn==1.1.3
scipy==1.9.3
threadpoolctl==3.1.0
tld==0.12.6
webencodings==0.5.1
Werkzeug==2.2.2
wincertstore==0.2
yarl==1.8.1
